backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Werkzeug==2.2.3
pyOpenSSL==23.2.0
